+++ test/edgeinfobuilder.cc
@@ -40,7 +40,6 @@ void TryWriteRead(const EdgeInfoBuilder& eibuilder) {
   }
 
   // validate the read in fields to the original EdgeInfoBuilder
-//  EdgeInfo* ei = &((reinterpret_cast<EdgeInfo*>(edgeinfo_ + offset))->SetPointers())
   EdgeInfo* ei = new EdgeInfo(memblock);
   ei->ToOstream();
 
@@ -75,7 +74,7 @@ void TryWriteRead(const EdgeInfoBuilder& eibuilder) {
 
 void TestWriteRead() {
   EdgeInfoBuilder eibuilder;
-  std::vector<size_t> street_name_offset_list;
+  std::vector<uint32_t> street_name_offset_list;
   street_name_offset_list.push_back(963);
   street_name_offset_list.push_back(957);
   street_name_offset_list.push_back(862);
